refactor: splite code into files
